+++ src/components/OrdersList/OrdersList.jsx
@@ -16,8 +16,6 @@ const OrdersList = () => {
 
   useEffect(() => {
     getOrders();
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return (
