+++ src/components/ProductsList/ProductsList.jsx
@@ -12,7 +12,6 @@ const ProductsList = () => {
 
   useEffect(() => {
     getProducts();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return (
