+++ src/hooks/useConfirm.js
@@ -1,11 +1,9 @@
-// eslint-disable-next-line import/prefer-default-export
 export const useConfirm = (
   message = "",
   fulfill = () => {},
   rejection = () => {}
 ) => {
   const confirmAction = () => {
-    // eslint-disable-next-line no-alert
     if (window.confirm(message)) {
       fulfill();
     } else {
