+++ src/store/index.js
@@ -1,7 +1,7 @@
 import { configureStore } from "@reduxjs/toolkit";
-import cartReducer from "./modules/cartSlice";
-import orderReducer from "./modules/orderSlice";
-import productReducer from "./modules/productSlice";
+import productReducer from "./modules/product/productSlice";
+import cartReducer from "./modules/cart/cartSlice";
+import orderReducer from "./modules/order/orderSlice";
 
 export default configureStore({
   reducer: { cart: cartReducer, order: orderReducer, product: productReducer },
